requirements.txt
PyPDF2==3.0.1
pdfplumber==0.10.3
pandas==2.1.4
google-generativeai==0.3.2
chromadb==0.4.22
sentence-transformers==2.2.2
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.3
python-dotenv==1.0.0
python-multipart==0.0.6
ipykernel==6.27.1
notebook==7.0.6